src/data/usecases/db-add-account.spec.ts
import { DbAddAccount } from './db-add-account'
import { mockAddAccountParams } from '@/domain/test/mocks/account'
import { HasherSpy } from '@/data/test/mocks/criptography'
import { AddAccountRepositorySpy } from '@/data/test/mocks/db-account'

type SutTypes = {
  sut: DbAddAccount
  hasherSpy: HasherSpy
  addAccountRepositoryStub: AddAccountRepositorySpy
}

const makeSut = (): SutTypes => {
  const hasherSpy = new HasherSpy()
  const addAccountRepositoryStub = new AddAccountRepositorySpy()
  const sut = new DbAddAccount(hasherSpy, addAccountRepositoryStub)
  return {
    sut,
    hasherSpy,
    addAccountRepositoryStub
  }
}

describe('DbAddAccount Usecases', () => {
  test('Should call Hasher with correct password', async () => {
    const { sut, hasherSpy } = makeSut()
    const addAccountParams = mockAddAccountParams()
    await sut.add(addAccountParams)
    expect(hasherSpy.plaintext).toBe(addAccountParams.password)
  })

  test('Should throw if Hasher throws', async () => {
    const { sut, hasherSpy } = makeSut()
    jest.spyOn(hasherSpy, 'hash').mockRejectedValueOnce(new Error())
    const addAccountParams = mockAddAccountParams()
    const promise = sut.add(addAccountParams)
    await expect(promise).rejects.toThrow()
  })

  test('Should call AddAccountRepository with correct values', async () => {
    const { sut, addAccountRepositoryStub, hasherSpy } = makeSut()
    const addAccountParams = mockAddAccountParams()
    await sut.add(addAccountParams)
    expect(addAccountRepositoryStub.addAccountParams).toEqual({
      name: addAccountParams.name,
      email: addAccountParams.email,
      password: hasherSpy.digest
    })
  })
})
